AutoJournAppFiles/backend/requirements.txt
accelerate==0.31.0
aniso8601==9.0.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.1.0
bertopic==0.16.2
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==0.29.37
filelock==3.15.4
Flask==3.0.3
flask-restx==1.3.0
fonttools==4.53.0
fsspec==2024.6.1
h11==0.14.0
hdbscan==0.8.37
huggingface-hub==0.23.4
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.22.0
praw==7.7.1
prawcore==2.4.0
preshed==3.0.9
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12.1
Telethon==1.36.0
tenacity==8.4.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
transformers==4.42.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
update-checker==0.18.0
urllib3==2.2.2
uvicorn==0.30.1
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
redis==4.5.5
langchain-ollama
langchain_community
keybert
#collections
networkx
openai
serpapi
readability-lxml
rank_bm25
bs4
newspaper3k
traceback2
rapidfuzz
gunicorn